[WIP] Re-do Attr using webidl2js
Missing pieces:
- TreatNullAs=EmptyString missing from webidl2js
- Need to wait for webidl2js to release instead of using master

This completely replaces the Attr implementation, and all related code. It is now based on the latest spec, in particular no longer inheriting from Node. It uses webidl2js to generate a lot of the boilerplate.

Along the way we update webidl2js to the latest version, requiring some minor changes in our code generation pipeline.

A couple other methods were updated with better IDL compliance in the process of getting tests to pass:
- Node.prototype.insertBefore
- Node.prototype.replaceChild
- Node.prototype.appendChild

A number of tests were removed since they test for things that are no longer true, like Attr having child nodes. Some of these that have changed results were introduced as new to-upstream web-platform-tests.